+++ docs/tethys_sdk/routing.rst
@@ -116,7 +116,7 @@ Bokeh Integration in Tethys takes advantage of :ref:`websockets` and ``Django Ch
 
 Even though Bokeh uses :ref:`websockets`, routing with Bokeh endpoints is handled differently from other :ref:`websockets` that would normally be handled by a ``Consumer`` class and use the :ref:`consumer-decorator`. In contrast, Bokeh uses a ``handler`` function that contains the main logic needed for a Bokeh model to be displayed. It contains the model or group of models as well as the callback functions that will help link them to the client. A ``handler`` function should be registered with the :ref:`handler-decorator`.
 
-``Handlers`` are added to the ``Bokeh Document``, the smallest serialization unit in ``Bokeh Server``. This same ``Document`` is later retrieved and added to the template variables in a ``controller`` function that will be linked to the ``Handler function`` using Bokeh's `server_document` function. The ``controller`` function will be created and registered automatically with the :ref:`handler-decorator`. However, you can manually create a ``controller`` fuction if custom logic is needed. In this case the ``controller`` function should not be decorated, but rather passed in as an argument to the :ref:`handler-decorator`.
+``Handlers`` are added to the ``Bokeh Document``, the smallest serialization unit in ``Bokeh Server``. This same ``Document`` is later retrieved and added to the template variables in a ``controller`` function that will be linked to the ``Handler function`` using Bokeh's `server_document` function. The ``controller`` function will be created and registered automatically with the :ref:`handler-decorator`. However, you can manually create a ``controller`` function if custom logic is needed. In this case the ``controller`` function should not be decorated, but rather passed in as an argument to the :ref:`handler-decorator`.
 
 A ``Bokeh Document`` comes with a ``Bokeh Request``. This request contains most of the common attributes of a normal ``HTTPRequest``, and can be easily converted to an ``HTTPRequest`` using the ``with_request`` argument in the :ref:`handler-decorator`. Similarly, the ``with_workspaces`` argument can be used to add ``user_workspace`` and ``app_workspace`` to the ``Bokeh Document``. This latter argument will also convert the ``Bokeh Request`` of the ``Document`` to an ``HTTPRequest``, meaning it will do the same thing as the ``with_request`` argument in addition to adding workspaces.
