+++ tethys_cli/scaffold_templates/app_templates/default/setup.py_tmpl
@@ -4,7 +4,7 @@ from tethys_apps.base.app_base import TethysAppBase
 
 # -- Apps Definition -- #
 app_package = '{{project}}'
-release_package = f'TethysAppBase.package_namespace-{app_package}'
+release_package = f'{TethysAppBase.package_namespace}-{app_package}'
 
 # -- Python Dependencies -- #
 dependencies = []
