+++ tethys_sdk/routing.py
@@ -8,7 +8,6 @@
 
 # flake8: noqa
 # DO NOT ERASE
-from tethys_apps.base.controller import TethysController, controller, consumer, handler
+from tethys_apps.base.controller import TethysController, controller, consumer, handler, register_controllers
 from tethys_apps.base.bokeh_handler import with_request, with_workspaces
 from tethys_apps.base.url_map import url_map_maker
-from tethys_apps.base.app_base import register_controllers
